+++ HISTORY.md
@@ -2,3 +2,8 @@
 
 ## v0.1.0 -- 2025/11/09
     - Basic Mechanics like main player and collision detection
+
+## v0.1.0 -- 2025/11/10
+    - completed basic implementation of enemy and player mechanics
+    - planes autoshoot mechancis
+    - plane death animation